+++ src/Connection/ConnectionInterface.php
@@ -14,7 +14,7 @@ use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Query\BatchQueryResultInterface;
 use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
-use Yiisoft\Db\Schema\Column\ColumnBuilder;
+use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
 use Yiisoft\Db\Schema\TableSchemaInterface;
@@ -87,11 +87,9 @@ interface ConnectionInterface
     public function close(): void;
 
     /**
-     * Returns the column builder class name for creating column instances.
-     *
-     * @psalm-return class-string<ColumnBuilder>
+     * Returns the column factory for creating column instances.
      */
-    public function getColumnBuilderClass(): string;
+    public function getColumnFactory(): ColumnFactoryInterface;
 
     /**
      * Returns the name of the DB driver for the current `dsn`.