+++ tests/AbstractColumnFactoryTest.php
@@ -17,7 +17,7 @@ abstract class AbstractColumnFactoryTest extends TestCase
     public function testFromDbType(string $dbType, string $expectedType, string $expectedInstanceOf): void
     {
         $db = $this->getConnection();
-        $columnFactory = $db->getColumnBuilderClass()::columnFactory();
+        $columnFactory = $db->getColumnFactory();
 
         $column = $columnFactory->fromDbType($dbType);
 
@@ -36,7 +36,7 @@ abstract class AbstractColumnFactoryTest extends TestCase
         array $expectedMethodResults = []
     ): void {
         $db = $this->getConnection();
-        $columnFactory = $db->getColumnBuilderClass()::columnFactory();
+        $columnFactory = $db->getColumnFactory();
 
         $column = $columnFactory->fromDefinition($definition);
 
@@ -61,7 +61,7 @@ abstract class AbstractColumnFactoryTest extends TestCase
         array $expectedMethodResults = []
     ): void {
         $db = $this->getConnection();
-        $columnFactory = $db->getColumnBuilderClass()::columnFactory();
+        $columnFactory = $db->getColumnFactory();
 
         $column = $columnFactory->fromPseudoType($pseudoType);
 
@@ -82,7 +82,7 @@ abstract class AbstractColumnFactoryTest extends TestCase
     public function testFromType(string $type, string $expectedType, string $expectedInstanceOf): void
     {
         $db = $this->getConnection();
-        $columnFactory = $db->getColumnBuilderClass()::columnFactory();
+        $columnFactory = $db->getColumnFactory();
 
         $column = $columnFactory->fromType($type);
 
@@ -93,7 +93,7 @@ abstract class AbstractColumnFactoryTest extends TestCase
     public function testFromDefinitionWithExtra(): void
     {
         $db = $this->getConnection();
-        $columnFactory = $db->getColumnBuilderClass()::columnFactory();
+        $columnFactory = $db->getColumnFactory();
 
         $column = $columnFactory->fromDefinition('char(1) NOT NULL', ['extra' => 'UNIQUE']);